ToolRunner is now somewhat less braindead, using different strategies for different commands. Commands executed during BuildModuleMenu() are blocking and run with minimum latency, while lengthy operations are executed asynchronously with timer-based polling and communication via wxEvents, similar to the gcc compiler plugin. The latter approach has some massive latency (300-400 ms), but since a commit can take 3 or 4 minutes, this actually matters very little. On the other hand, the message-based approach is a lot more suitable when running long transactions over a slow network, as it allows the user to work on, and the application does not freeze. Note, however, that this means the user can modify and save files or even close code::blocks while a commit is running(!). The plugin tries to at least partially catch this problem by calling Finish() in its OnModuleUnload() function. As transactions are asynchronous now, conditional actions (like update after commit / update on out-of-date) are currently dysfunct and need to be rewritten using message passing yet.
The paths to TortoiseProc and TortoiseAct are no longer saved to the configuration. If either of the two Tortoises is installed, then they will be found through the registry anyway.
There is also no more NastyFind() on TortoiseProc and TortoiseAct. There is little point in searching for them if installation was not properly done.
Instead of the path to TortoiseProc, the preferences dialog now allows the setting of the cvs binary path.
Pre-compilation steps now use test -nt and upx is run in post-compilation steps.