Implemented 'svn export' both with known destination and with automatic tempfile.
DIFF support works again, and it is idependent of either Windows or TortoiseSVN:
- On Windows, if present, TortoiseSVN will be instructed to call whatever the user has configured to view DIFF
  - by default, this is TortoiseMerge
  - if TortoiseSVN is not present, KDiff3 will be used
  - if KDiff3 is not found either, then WinMerge will be used
- On Linux, KDiff3 is used as the default
  - if KDiff3 is not found, tkdiff.tcl is used
  - tkdiff.tcl (as the name suggests) is TK-based, therefore it runs on practically every machine

Similarly, conflict editing works with different tools now.
'svn resolved' command implemented.

Changed cutoff time for tempfiles to 4 hours (it is unlikely anyone will notice the 20 or 30 tempfiles accumulated during that time).
Added missing zero initialisation of implicit_run to ToolRunner.
OnFirstRun() is no longer just run on the first run, and is therefore called SearchBinaries(). It searches for the svn and cvs binaries as well as for external diff applications.